Change class and package names
Authored-by: Aaron Lindsey <alindsey@pivotal.io>
