+++ geode-assembly/src/acceptanceTest/java/org/apache/geode/metrics/function/executions/ExecutionsTimerValues.java
@@ -12,7 +12,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.apache.geode.metrics.functionexecutions;
+package org.apache.geode.metrics.function.executions;
 
 import java.io.Serializable;
